Factor out i18n-base.ts.
