+++ client/src/i18n-base.test.ts
@@ -1,0 +1,6 @@
+import { isSupportedLocale } from "./i18n-base";
+
+test("isSupportedLocale() works", () => {
+  expect(isSupportedLocale('en')).toBe(true);
+  expect(isSupportedLocale('zz')).toBe(false);
+});
